Improving modularization
